Recette - Corrections diverses (#3086)
* Fixes

* Linting
