+++ package.json
@@ -51,7 +51,9 @@
     "@opentelemetry/context-async-hooks": "^1.19.0",
     "@opentelemetry/exporter-trace-otlp-http": "^0.46.0",
     "@opentelemetry/instrumentation": "^0.41.0",
+    "@opentelemetry/resources": "^1.21.0",
     "@opentelemetry/sdk-trace-base": "^1.19.0",
+    "@opentelemetry/semantic-conventions": "^1.21.0",
     "@prisma/client": "5.7.1",
     "@prisma/instrumentation": "5.7.1",
     "@sentry/browser": "^6.19.7",
@@ -268,9 +270,9 @@
     "jest-environment-node": "^29.4.1",
     "jest-transform-stub": "^2.0.0",
     "jsdom": "~22.1.0",
+    "mermaid": "10.6.1",
     "node-mocks-http": "^1.11.0",
     "nx": "17.0.1",
-    "mermaid": "10.6.1",
     "prettier": "^2.8.6",
     "raw-loader": "^4.0.2",
     "sass": "^1.55.0",
